+++ commands/ping.js
@@ -1,0 +1,10 @@
+import { SlashCommandBuilder } from 'discord.js';
+
+export default {
+	data: new SlashCommandBuilder()
+		.setName('ping')
+		.setDescription('Replies with Pong!'),
+	async execute(interaction) {
+		await interaction.reply('Pong!');
+	},
+};
